+++ core/src/OpenMP/Kokkos_OpenMP_MDRangePolicy.hpp
@@ -42,33 +42,19 @@
 //@HEADER
 */
 
-#ifndef KOKKOS_OPENACC_MDRANGE_POLICY_HPP_
-#define KOKKOS_OPENACC_MDRANGE_POLICY_HPP_
+#ifndef KOKKOS_OPENMP_MDRANGEPOLICY_HPP_
+#define KOKKOS_OPENMP_MDRANGEPOLICY_HPP_
 
 #include <KokkosExp_MDRangePolicy.hpp>
 
-template <>
-struct Kokkos::default_outer_direction<Kokkos::Experimental::OpenACC> {
-  using type                     = Iterate;
-  static constexpr Iterate value = Iterate::Left;
-};
-
-template <>
-struct Kokkos::default_inner_direction<Kokkos::Experimental::OpenACC> {
-  using type                     = Iterate;
-  static constexpr Iterate value = Iterate::Left;
-};
-
 namespace Kokkos {
-namespace Experimental {
 namespace Impl {
 
+// Settings for TeamMDRangePolicy
 template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, OpenACC, ThreadAndVector>
-    : AcceleratorBasedNestLevel<Rank, ThreadAndVector> {};
+struct ThreadAndVectorNestLevel<Rank, OpenMP, ThreadAndVector>
+    : HostBasedNestLevel<Rank, ThreadAndVector> {};
 
 }  // namespace Impl
-}  // namespace Experimental
 }  // namespace Kokkos
-
 #endif
