+++ core/src/decl/Kokkos_Declare_SERIAL.hpp
@@ -47,6 +47,7 @@
 
 #if defined(KOKKOS_ENABLE_SERIAL)
 #include <Kokkos_Serial.hpp>
+#include <Serial/Kokkos_Serial_MDRangePolicy.hpp>
 #endif
 
 #endif
