+++ core/src/decl/Kokkos_Declare_THREADS.hpp
@@ -47,6 +47,7 @@
 
 #if defined(KOKKOS_ENABLE_THREADS)
 #include <Kokkos_Threads.hpp>
+#include <Threads/Kokkos_Threads_MDRangePolicy.hpp>
 #endif
 
 #endif
